requirements.txt
Django==2.0.2
django_localflavor==2.0